+++ src/ts/AdMob/Views/AdMobView.ts
@@ -67,6 +67,7 @@ export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandl
     private _gdprPopupClicked: boolean = false;
     private _admobOMController: AdmobOpenMeasurementController | undefined;
     private _deviceInfo: DeviceInfo;
+    private _volume: number = 1;
 
     constructor(platform: Platform, core: ICoreApi, adMobSignalFactory: AdMobSignalFactory, container: AdUnitContainer, campaign: AdMobCampaign, deviceInfo: DeviceInfo, gameId: string, privacy: AbstractPrivacy, showGDPRBanner: boolean, om: AdmobOpenMeasurementController | undefined) {
         super(platform, 'admob');
@@ -94,7 +95,8 @@ export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandl
             onAFMAResolveOpenableIntents: (request) => this.onResolveOpenableIntents(request),
             onAFMATrackingEvent: (event, data?) => this.onTrackingEvent(event, data),
             onAFMAClickSignalRequest: (touchInfo) => this.onClickSignalRequest(touchInfo),
-            onAFMAUserSeeked: () => this.onUserSeeked()
+            onAFMAUserSeeked: () => this.onUserSeeked(),
+            onVolumeChange: (volume) => { this._volume = volume; }
         });
         this._mraidBridge = new MRAIDBridge(core, {
             onSetOrientationProperties: (allowOrientation: boolean, forceOrientation: Orientation) => this.onSetOrientationProperties(allowOrientation, forceOrientation)
@@ -135,6 +137,10 @@ export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandl
         this._deviceInfo.checkIsMuted();
 
         this.choosePrivacyShown();
+    }
+
+    public getVideoPlayerVolume(): number {
+        return this._volume;
     }
 
     public hide() {